+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/constants.go
@@ -10,8 +10,42 @@ package armsqlvirtualmachine
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine"
-	moduleVersion = "v0.10.0"
+	moduleVersion = "v0.11.0"
 )
+
+// AdditionalOsPatch - Additional VM Patching solution enabled on the Virtual Machine
+type AdditionalOsPatch string
+
+const (
+	AdditionalOsPatchWSUS AdditionalOsPatch = "WSUS"
+	AdditionalOsPatchWU   AdditionalOsPatch = "WU"
+	AdditionalOsPatchWUMU AdditionalOsPatch = "WUMU"
+)
+
+// PossibleAdditionalOsPatchValues returns the possible values for the AdditionalOsPatch const type.
+func PossibleAdditionalOsPatchValues() []AdditionalOsPatch {
+	return []AdditionalOsPatch{
+		AdditionalOsPatchWSUS,
+		AdditionalOsPatchWU,
+		AdditionalOsPatchWUMU,
+	}
+}
+
+// AdditionalVMPatch - Additional Patch to be enable or enabled on the SQL Virtual Machine.
+type AdditionalVMPatch string
+
+const (
+	AdditionalVMPatchMicrosoftUpdate AdditionalVMPatch = "MicrosoftUpdate"
+	AdditionalVMPatchNotSet          AdditionalVMPatch = "NotSet"
+)
+
+// PossibleAdditionalVMPatchValues returns the possible values for the AdditionalVMPatch const type.
+func PossibleAdditionalVMPatchValues() []AdditionalVMPatch {
+	return []AdditionalVMPatch{
+		AdditionalVMPatchMicrosoftUpdate,
+		AdditionalVMPatchNotSet,
+	}
+}
 
 // AssessmentDayOfWeek - Day of the week to run assessment.
 type AssessmentDayOfWeek string
@@ -129,15 +163,15 @@ func PossibleClusterSubnetTypeValues() []ClusterSubnetType {
 type Commit string
 
 const (
-	CommitASYNCHRONOUSCOMMIT Commit = "ASYNCHRONOUS_COMMIT"
-	CommitSYNCHRONOUSCOMMIT  Commit = "SYNCHRONOUS_COMMIT"
+	CommitAsynchronousCommit Commit = "Asynchronous_Commit"
+	CommitSynchronousCommit  Commit = "Synchronous_Commit"
 )
 
 // PossibleCommitValues returns the possible values for the Commit const type.
 func PossibleCommitValues() []Commit {
 	return []Commit{
-		CommitASYNCHRONOUSCOMMIT,
-		CommitSYNCHRONOUSCOMMIT,
+		CommitAsynchronousCommit,
+		CommitSynchronousCommit,
 	}
 }
 
@@ -229,15 +263,15 @@ func PossibleDiskConfigurationTypeValues() []DiskConfigurationType {
 type Failover string
 
 const (
-	FailoverAUTOMATIC Failover = "AUTOMATIC"
-	FailoverMANUAL    Failover = "MANUAL"
+	FailoverAutomatic Failover = "Automatic"
+	FailoverManual    Failover = "Manual"
 )
 
 // PossibleFailoverValues returns the possible values for the Failover const type.
 func PossibleFailoverValues() []Failover {
 	return []Failover{
-		FailoverAUTOMATIC,
-		FailoverMANUAL,
+		FailoverAutomatic,
+		FailoverManual,
 	}
 }
 
@@ -310,17 +344,17 @@ func PossibleOperationOriginValues() []OperationOrigin {
 type ReadableSecondary string
 
 const (
-	ReadableSecondaryALL      ReadableSecondary = "ALL"
-	ReadableSecondaryNO       ReadableSecondary = "NO"
-	ReadableSecondaryREADONLY ReadableSecondary = "READ_ONLY"
+	ReadableSecondaryAll      ReadableSecondary = "All"
+	ReadableSecondaryNo       ReadableSecondary = "No"
+	ReadableSecondaryReadOnly ReadableSecondary = "Read_Only"
 )
 
 // PossibleReadableSecondaryValues returns the possible values for the ReadableSecondary const type.
 func PossibleReadableSecondaryValues() []ReadableSecondary {
 	return []ReadableSecondary{
-		ReadableSecondaryALL,
-		ReadableSecondaryNO,
-		ReadableSecondaryREADONLY,
+		ReadableSecondaryAll,
+		ReadableSecondaryNo,
+		ReadableSecondaryReadOnly,
 	}
 }
 
@@ -328,15 +362,15 @@ func PossibleReadableSecondaryValues() []ReadableSecondary {
 type Role string
 
 const (
-	RolePRIMARY   Role = "PRIMARY"
-	RoleSECONDARY Role = "SECONDARY"
+	RolePrimary   Role = "Primary"
+	RoleSecondary Role = "Secondary"
 )
 
 // PossibleRoleValues returns the possible values for the Role const type.
 func PossibleRoleValues() []Role {
 	return []Role{
-		RolePRIMARY,
-		RoleSECONDARY,
+		RolePrimary,
+		RoleSecondary,
 	}
 }
 
@@ -362,7 +396,8 @@ func PossibleSQLImageSKUValues() []SQLImageSKU {
 	}
 }
 
-// SQLManagementMode - SQL Server Management type.
+// SQLManagementMode - SQL Server Management type. NOTE: This parameter is not used anymore. API will automatically detect
+// the Sql Management, refrain from using it.
 type SQLManagementMode string
 
 const (
@@ -475,5 +510,23 @@ const (
 func PossibleTroubleshootingScenarioValues() []TroubleshootingScenario {
 	return []TroubleshootingScenario{
 		TroubleshootingScenarioUnhealthyReplica,
+	}
+}
+
+// VMIdentityType - Identity type of the virtual machine. Specify None to opt-out of Managed Identities.
+type VMIdentityType string
+
+const (
+	VMIdentityTypeNone           VMIdentityType = "None"
+	VMIdentityTypeSystemAssigned VMIdentityType = "SystemAssigned"
+	VMIdentityTypeUserAssigned   VMIdentityType = "UserAssigned"
+)
+
+// PossibleVMIdentityTypeValues returns the possible values for the VMIdentityType const type.
+func PossibleVMIdentityTypeValues() []VMIdentityType {
+	return []VMIdentityType{
+		VMIdentityTypeNone,
+		VMIdentityTypeSystemAssigned,
+		VMIdentityTypeUserAssigned,
 	}
 }