+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/models.go
@@ -106,6 +106,9 @@ type AutoBackupSettings struct {
 
 // AutoPatchingSettings - Set a patching window during which Windows and SQL patches will be applied.
 type AutoPatchingSettings struct {
+	// Additional Patch to be enable or enabled on the SQL Virtual Machine.
+	AdditionalVMPatch *AdditionalVMPatch
+
 	// Day of week to apply the patch on.
 	DayOfWeek *DayOfWeek
 
@@ -338,6 +341,9 @@ type PrivateIPAddress struct {
 
 // Properties - The SQL virtual machine properties.
 type Properties struct {
+	// Additional VM Patching solution enabled on the Virtual Machine
+	AdditionalVMPatch *AdditionalOsPatch
+
 	// SQL best practices Assessment Settings.
 	AssessmentSettings *AssessmentSettings
 
@@ -362,7 +368,8 @@ type Properties struct {
 	// SQL Server edition type.
 	SQLImageSKU *SQLImageSKU
 
-	// SQL Server Management type.
+	// SQL Server Management type. NOTE: This parameter is not used anymore. API will automatically detect the Sql Management,
+	// refrain from using it.
 	SQLManagement *SQLManagementMode
 
 	// SQL Server license type.
@@ -376,6 +383,9 @@ type Properties struct {
 
 	// Storage Configuration Settings.
 	StorageConfigurationSettings *StorageConfigurationSettings
+
+	// Virtual Machine Identity details used for Sql IaaS extension configurations.
+	VirtualMachineIdentitySettings *VirtualMachineIdentity
 
 	// ARM Resource id of underlying virtual machine created from SQL marketplace image.
 	VirtualMachineResourceID *string
@@ -452,6 +462,9 @@ type SQLStorageSettings struct {
 
 	// Logical Unit Numbers for the disks.
 	Luns []*int32
+
+	// Use storage pool to build a drive if true or not provided
+	UseStoragePool *bool
 }
 
 // SQLStorageUpdateSettings - Set disk storage settings for SQL Server.
@@ -494,6 +507,9 @@ type SQLTempDbSettings struct {
 
 	// SQL Server tempdb persist folder location
 	PersistFolderPath *string
+
+	// Use storage pool to build a drive if true or not provided
+	UseStoragePool *bool
 }
 
 // SQLVMTroubleshooting - Details required for SQL VM troubleshooting
@@ -519,7 +535,7 @@ type SQLVirtualMachine struct {
 	// REQUIRED; Resource location.
 	Location *string
 
-	// Azure Active Directory identity of the server.
+	// DO NOT USE. This value will be deprecated. Azure Active Directory identity of the server.
 	Identity *ResourceIdentity
 
 	// Resource properties.
@@ -591,6 +607,9 @@ type ServerConfigurationsManagementSettings struct {
 type StorageConfigurationSettings struct {
 	// Disk configuration to apply to SQL Server.
 	DiskConfigurationType *DiskConfigurationType
+
+	// Enable SQL IaaS Agent storage configuration blade in Azure Portal.
+	EnableStorageConfigBlade *bool
 
 	// SQL Server Data Storage Settings.
 	SQLDataSettings *SQLStorageSettings
@@ -664,8 +683,20 @@ type UnhealthyReplicaInfo struct {
 
 // Update - An update to a SQL virtual machine.
 type Update struct {
+	// Resource properties.
+	Properties *Properties
+
 	// Resource tags.
 	Tags map[string]*string
+}
+
+// VirtualMachineIdentity - Virtual Machine Identity details used for Sql IaaS extension configurations.
+type VirtualMachineIdentity struct {
+	// Resource Id of the identity. Only required when userAssigned identity is selected.
+	ResourceID *string
+
+	// Identity type of the virtual machine. Specify None to opt-out of Managed Identities.
+	Type *VMIdentityType
 }
 
 // WsfcDomainCredentials - Domain credentials for setting up Windows Server Failover Cluster for SQL availability group.
@@ -697,6 +728,9 @@ type WsfcDomainProfile struct {
 
 	// Optional path for fileshare witness.
 	FileShareWitnessPath *string
+
+	// The flag to check if SQL service account is GMSA.
+	IsSQLServiceAccountGmsa *bool
 
 	// Organizational Unit path in which the nodes and cluster will be present.
 	OuPath *string