+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Release History
 
+## 0.11.0 (2023-12-09)
+### Breaking Changes
+
+- `CommitASYNCHRONOUSCOMMIT`, `CommitSYNCHRONOUSCOMMIT` from enum `Commit` has been removed
+- `FailoverAUTOMATIC`, `FailoverMANUAL` from enum `Failover` has been removed
+- `ReadableSecondaryALL`, `ReadableSecondaryNO`, `ReadableSecondaryREADONLY` from enum `ReadableSecondary` has been removed
+- `RolePRIMARY`, `RoleSECONDARY` from enum `Role` has been removed
+
+### Features Added
+
+- New value `CommitAsynchronousCommit`, `CommitSynchronousCommit` added to enum type `Commit`
+- New value `FailoverAutomatic`, `FailoverManual` added to enum type `Failover`
+- New value `ReadableSecondaryAll`, `ReadableSecondaryNo`, `ReadableSecondaryReadOnly` added to enum type `ReadableSecondary`
+- New value `RolePrimary`, `RoleSecondary` added to enum type `Role`
+- New enum type `AdditionalOsPatch` with values `AdditionalOsPatchWSUS`, `AdditionalOsPatchWU`, `AdditionalOsPatchWUMU`
+- New enum type `AdditionalVMPatch` with values `AdditionalVMPatchMicrosoftUpdate`, `AdditionalVMPatchNotSet`
+- New enum type `VMIdentityType` with values `VMIdentityTypeNone`, `VMIdentityTypeSystemAssigned`, `VMIdentityTypeUserAssigned`
+- New struct `VirtualMachineIdentity`
+- New field `AdditionalVMPatch` in struct `AutoPatchingSettings`
+- New field `AdditionalVMPatch`, `VirtualMachineIdentitySettings` in struct `Properties`
+- New field `UseStoragePool` in struct `SQLStorageSettings`
+- New field `UseStoragePool` in struct `SQLTempDbSettings`
+- New field `EnableStorageConfigBlade` in struct `StorageConfigurationSettings`
+- New field `Properties` in struct `Update`
+- New field `IsSQLServiceAccountGmsa` in struct `WsfcDomainProfile`
+
+
 ## 0.10.0 (2023-11-24)
 ### Features Added